requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.4
django-appconf==1.0.4
django-import-export==2.5.0
django-redis==4.12.1
django-rest-framework==0.1.0
django-select2==7.7.1
djangorestframework==3.12.4
djangorestframework-datatables==0.7.0
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
psycopg2-binary==2.9.3
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
